requirements/base.txt
Django==5.0.4
djangorestframework==3.14.0
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.27.1
django-filter==21.1
celery==5.2.3
redis==4.2.2
python-dotenv==0.19.2
gunicorn==20.1.0
africastalking==1.2.9
whitenoise==6.6.0
django-mptt==0.14.0